Conflicts:
	arch/amd64/amd64/mainbus.c
	arch/amd64/pci/iommu.c
	arch/amd64/pci/pci_machdep.c
	arch/i386/pci/pci_machdep.c
	dev/pci/pci.c
	kern/subr_extent.c
	sys/extent.h

+++ arch/sparc/dev/qec.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: qec.c,v 1.17 2006/06/02 20:00:54 miod Exp $	*/
+/*	$OpenBSD: qec.c,v 1.19 2010/09/05 18:10:10 kettenis Exp $	*/
 
 /*
  * Copyright (c) 1998 Theo de Raadt and Jason L. Wright.
@@ -35,7 +35,6 @@
 #include <sys/malloc.h>
 #include <sys/buf.h>
 #include <sys/proc.h>
-#include <sys/user.h>
 #include <sys/mbuf.h>
 #include <sys/socket.h>
 
@@ -249,9 +248,9 @@ qec_translate(sc, ca)
 
 		for (j = 0; j < sc->sc_nrange; j++) {
 			if (sc->sc_range[j].cspace == cspace) {
-				(int)ca->ca_ra.ra_reg[i].rr_paddr +=
+				ca->ca_ra.ra_reg[i].rr_paddr +=
 					sc->sc_range[j].poffset;
-				(int)ca->ca_ra.ra_reg[i].rr_iospace =
+				ca->ca_ra.ra_reg[i].rr_iospace =
 					sc->sc_range[j].pspace;
 				break;
 			}
